Propose a way to add validation on the enum attribute
Last year, I heard about the inconveniences of ActiveRecord Enum and got
some feedback (one of those is #40456).

Some people have reported that in practice they are using enums for user
input and it is inconvenient to not be able to handle invalid input with
validation like other attributes.

Actually, a former colleague was overriding the attribute writer for
validation, and they had wished this inconvenience to be resolved in the
Rails.

So I propose a way to add validation on the enum attribute. This allows
people to no longer need to override the attribute writer.

Before:

```ruby
class Book < ActiveRecord::Base
  enum :status, [:proposed, :written]

  validates_inclusion_of :status, in: statuses.keys

  def status=(value)
    super
  rescue ArgumentError
    @attributes.write_cast_value("status", value)
  end
end
```

After:

```ruby
class Book < ActiveRecord::Base
  enum :status, [:proposed, :written], validate: true
end
```

Resolves #13971.