Add table for shallow routing to routing doc
Shallow routes is one of the most difficult-to-use features for
generating routes in Rails.
To make it easier to understand how it works, a routing table
shows us the final result and really helps.
However, there's only one table for shallow routes and it's with
shallow_path option, which is even more complicated.
Adding a table for simpler shallow routes will help people
understand shallow routing's idea and usage better.

There's a drawback of this addition. Adding a large table occupies
the screen and might make it slightly difficult to walk through
the documentation.