Stop executing perform method when retry_job is called
While `retry_job` would normally be expected to be called by the framework
itself (in response to `retry_on` being configured for example) or possibly by
a library like `job-iteration` that breaks up jobs into smaller units of work,
it is possible that a job itself could choose to call `retry_job`. If this
call to `retry_job` is not the final statement executed by the job, the job
would enter a strange state where it has been re-enqueued for retry and also
continues to execute. The remainder of that execution could result in an
exception being raised that also could cause the job to be re-enqueued.

It seems like this is generally undesirable and unexpected.

This PR adds the ability to configure a job to abort the execution of its
`perform` method when `retry_job` is called (via raising an exception that
is caught and ignored one level above the `perform` call):

    # typed: true
    # frozen_string_literal: true
    class HaltRetryJob < ActiveJob::Base
      queue_as :low
      abort_perform_on_retry true

      def perform(params)
        execute_some_code

        retry_job

        # this will not be called if `abort_perform_on_retry true` is set
        some_other_code
      end
    end

Perhaps after a deprecation cycle, aborting the execution of the `perform`
method could become the default option rather than something that is opted-in
to.