More flexible action_on_unpermitted_parameters configuration
This PR gives more control over the behavior of Rails when an unpermitted parameter is encountered. The primary use case for this is to allow for synchronous use of the Rails logger; currently the :log option does not provide enough details about the controller to make the events emitted via ActiveSupport::Notifications actionable. While this could be modified, for backwards compatiability and ultimate flexiblity, we can simply allow the user to specify the behavior they desire.