Help identifying why we have an InvalidAuthenticityToken exception
We want to be able to more easily identify if it is a
valid_request_origin validation error or a HTTP Origin header mismatch.

Previously we had too look at the logs to know which type of issue it
was.

Related:
  * https://discuss.rubyonrails.org/t/proposal-for-improving-invalidauthenticitytoken-error-when-invalid-same-origin/77509
  * https://github.com/rails/rails/issues/38447

Extract warning_message assignment to a method

Avoid having to know the ProtectionMethods::Exception interface

Properly scope exception test in exception test class

There is no need to match error raised every times

Use clearer test description