Use stable branch for --edge option when not alpha
Follow-up to #41568.

If a user runs `rails new --edge` with an `x.y.0.betaN` or `x.y.z.rcN`
version, they most likely want to track the `x-y-stable` branch rather
than `main` branch.  In some cases, the `x-y-stable` branch may not
exist yet.  In those cases, it is better to raise an error when `bundle
install` fails, so that the user knows that the branch is not available.