Extend example code for autoloading serializers
Performing autoloading during initialization is not supported by Rails and will result
in a detailed deprecation warning. Following the proposed fix in the deprecation warning
will not work with the given piece of code however.

According to discussions in https://github.com/rails/rails/issues/37690 and
https://github.com/rails/rails/pull/41849, the example code has been amended to
ensure that serializers will only be loaded once.