Stop checking for Process.pid and trust the `fork` decorators
`Process.pid` is not cached on Linux and is surprisingly
expensive. These checks are in hotspots, the the cost is not
negligible.

In https://github.com/rails/rails/pull/37296#discussion_r330222303
we established that it's pretty much impossible to bypass
`Process.fork` / `Kernel#fork`.

We still added the checks to protect against user defined
`after_fork` hooks that would use the connection. But
I don't think the cost is worth this extra safety.