Turn ActiveSupport::TimeWithZone into a subclass of Time
When this feature was introduced it could either wrap a DateTime or a
Time instance due to the limited range that Time supported in Ruby 1.8.
Now that Time takes a timezone object and Rails 7.0 drops support for
Ruby 2.5 we can refactor it to be a proper subclass and avoid duck-typing
hacks like overriding `name` and `is_a?`.