Improve did_you_mean partial name correction
Before this change did_you_mean shows partial paths like `animals/_partial`,
but adding that to your render call in a view like `<%= render 'animals/_partial' %>`
 will still be missing as rails will search for the template `animals/__partial`.
We can provide the user a easier copy/paste correction if we don't tell them about the underscore.

This also reduces the candidates to only return partials if you are
looking for a partial and only look for non-templates when you are not
looking for a template.