Generate the `Permissions-Policy` header
This changes the `ActionDispatch::PermissionsPolicy` to generate both
the `Permissions-Policy` header and the `Feature-Policy` header.

Both headers are generated because most current browsers don't support
`Permissions-Policy` yet.

Furthermore the list of supported directives now includes `interest-cohort`
though it is not a part of the list of available directives from W3C.
This is to enable opting out of FLoC (https://web.dev/floc/#how-can-websites-opt-out-of-the-floc-computation).