Expand path of user provided file in runner
If user provides a file without absolute or relative path and
the file exists in any of the directories of the $LOAD_PATH
the `Kernel.load` method will load the latter one instead of the
intended one.

Expanding the path of the file forces Kernel to load the
correct file using its absolute path.

* Fixes #42007