Add batched_method for custom batch preloading in ActiveRecord
Often with Rails models, we end up iterating over a collection or
relation and calling the same method repeatedly. For these cases,
`includes` comes in really handy, but it has the limitation that it
only works on associations. More complex logic needs separate one-off
code to load in a way that will still avoid N+1s.

We write a lot of code like this in our applications, and it became
clear that a lot of these shared the same type of interface. There would
be a class method that took in an `Array` of objects to load, and
returned a `Hash` where the keys of the hash are the elements in the
`Array` and the values are the results, which would be loaded all at once.
What would it look like to build this pattern into Rails?

This commit proposes an interface called `batched_method` which is a DSL
allowing for the definition of methods in models which play nice
with `includes` & `preload`, but can have any definition. The idea is
that you'd define a batch method like:

``` ruby
class Post < ActiveRecord::Base
  # An implementation which takes in an Array[Post], and returns
  # a Hash[Post]=>Array[Comment]
  batch_method(:featured_comments) do |posts|
    comments_by_post_id = Comment.featured.where(post: posts).group_by(&:post_id)
    posts.index_with { |post| comments_by_post_id[post.id] }
  end
end
```

By writing the above implementation, you get the ability to call
`featured_comments` on a `Post` object like a normal method, but you can
also use `includes(:featured_comments)` and it loads them all the first
time that a single one is accessed (!).

I've been playing with this idea over in https://github.com/seejohnrun/prelude
(see README for more detail on motivations) for a bit and really think that
this would be a nice addition to Rails to help application authors think more
in terms of batching when iterating over model objects. I've taken
inspiration from over there, but rewritten the implementation to be clearer.

There are a bunch of ways that this can get more robust in the future,
but for this commit I've included a few neat additional features:

* The ability to define a `batch_size` on individual batch method
  definitions so that they load elements in batches only up to a certain
  size.

* The ability to define batched methods that take arguments. This is
  useful for common patterns like authentication checks where you want
  to batch things like `post.editable_by?(current_user)`.

* Compatibility with batch methods that happen to return `Hash#default_proc`.

Thanks for reading and I appreciate any feedback!