Refactor LocalCache to avoid calling Marshal.dump as much (take 2)
The local cache need to protect against mutations of both
the initial received value, and the value returned.

See:

  - https://github.com/rails/rails/pull/36656
  - https://github.com/rails/rails/pull/37587

Because of this, the overhead of the local cache end up being
quite significant. On a single read, the value will be deep duped
twice. So unless the value is one of the type benefiting from a
fast path, we'll have to do two `Marshal.load(Marshal.dump(value))`
roundtrips, which for large values can be very expensive.

By using a specialized `LocalEntry` type instead, we can store
the `Marshal` payload rather than the original value. This way
the overhead is reduced to a single `Marshal.dump` on writes
and a single `Marshal.load` on reads.