Make ModelSchema.inheritance_column a class_attribute
Ultimately it's exactly the same semantic that was implemented
using `superclass`, except it avoid recursion and benefit from
method cache, so much faster the longer the inheritance chain.

The use of an alias_method chain is unfortunate, but I don't
see any alternative to cast the provided value to string.