Fix validation of bi-directional accepts_nested_attributes
When defining bi-directional `accepts_nested_attributes` the validation
of a collection doesn't work well: in particular only the last record
of the collection is used. This happens because each child item in
the collection calls `valid?` on the parent record erasing the previous
validation errors stored in the parent.

Fixes the issue by tracking the record under validation and
skipping cyclic `valid?` calls.

Fixes #41811