Setup autoload on SchemaCache
Since b71c1cddddea736ae5af70474fe9bd765596e727, the SchemaCache
is loaded during the application initialization before the connection
adapter is loaded, which means that the application could fail to boot.

To fix this issue an autoload is setup for the schema cache and the
direct require is removed.

This autoload doesn't need to be moved to `eager_autoload` block
because the constant is already referenced in an initializer.