Add documentation around nested_attributes edge case
When nested_attributes, and collection ids are passed in the same update
or create, order impacts the outcome of the call.

If the collection ids are passed first, then the nested_attributes
update those ids, and create any records without ids that were passed.

If the nested_attributes are passed first, they are quashed by the ids
being passed.