make Rails::Rack::Logger tag compute configurable
For applications that use a structured/JSON logger, this is the first
change necessary to allow them to use the Tagged Logging feature in
Rails::Rack::Logger. Instead of returning an array of strings, a
configured tag computer could return a hash of fields to be merged into
a structured log.