Expect blocked host to be shown in the error
In both test cases I rewrote expected behavior to expect the blocked host to be
shown instead of the one allowed.

The problem is that we show `request.host` in the error message which is always
equal to `X-Forwarded-Host` if it's present.

This could be pretty confusing with valid nginx proxy configuration:

	location /api {
		proxy_set_header X-Forwarded-Host $host;
		proxy_pass http://api;
	}

With this configuration, the `Host` header will be equal to nginx' `$proxy_host`
by default, which equals "api" in that case. And the error tells developers
"Blocked host: localhost" which is very confusing since localhost is allowed by
default in development. The first intention is to try to add "localhost" to the
allowed hosts `config.hosts << "localhost"` which doesn't help.