Special case SafeBuffer#concat(nil) to avoid the bulk of deprecations
After trying out https://github.com/rails/rails/pull/42123
it appears that the bulk of the deprecations are for cases such as:

```ruby
output << some_helper
```

Where `some_helper` happens to sometime return `nil`.

While a regular String would indeed raise a `TypeError`
on such case, I wonder if it wouldn't be worth to special
case this specific scenario as it is quite harmless and would
drastically reduce the amount of deprecated code.