Update docs for ActiveRecord::Middleware::DatabaseSelector [ci skip]
The docs for this middleware indicate that you can set a delay that
can be a duration (eg. `2.seconds`).

https://api.rubyonrails.org/v6.1/classes/ActiveRecord/Middleware/DatabaseSelector.html

In normal apps, this works fine without an explicit require because
the extension is required through other libs. However, when creating a
minimal app, this results in an undefined method error.

This is something that is already being included in
test/development/production config files since c8c5497.