Replace EMPTY_ARRAY with literal
EMPTY_ARRAY is not referenced from any other files, and used only in one
place a few lines below.
This constant is not useful in terms of readability since `[].freeze` is
just readable as `empty array`, meaning we don't have to name it
explicitly.

This constant has been introduced in 0740adb379e97d873e024972bd6d07caf8801201
and the main purpose seems to be performance improvement.