Handle duplicate migration names in multi db
In a multidatabse setup, migrations live in seperate directories. The check for duplicate migration names, does not take this into
account.
So you __can__ end up having multiple migration with the same name.

When running `rails db:migrate` from the console, this other migration
might be loaded __before__ running the migrations for "this" database.

Because they have the same name, we have to ensure that the
contents of this migration is not carried over from "the other"
migrations.